Fix inline tag with code
The cause of this problem is in how inline tags work: you encounter #[,
then you take the rest of the line and feeds it into a new lexer. Using
a special option `this.interpolated` and a flag `this.ended`, the child
lexer returns whatever it can lex (i.e. the interpolated part) and the
rest (after the interpolation) is fed back to the parent lexer from
`child.input`.

For text, it works well, since can't really have ] as a valid part of
the text, so one can just use indexOf(']') for the ending index of the
interpolation.

Right now for code, it just feeds the **entire** rest of the line to the
token, which in many cases will not be valid JS and caught in the
assertExpression below. Plus, with code one never knows when code starts
or ends, if the `']'` is in quotes or not. So one has to be cautious
when consuming the strings, not using the crude `indexOf` check, which
is what I am doing here.

This also fixes jadejs/jade#1871.