Emit interpolated-code token
Previously, text that contains interpolated code was not parsed, but
rather left for jade-code-gen to handle. With the Jade AST in mind,
this is confusing since Text nodes may contain Code.

A separate token type is needed since `code` tokens are of a different
nature: `code` tokens need to be indented one more level from the tag
while `interpolated-code` tokens are inlined and can be adjacent to
text.